Move `PolarisVersionProvider` to `:polaris-version` to make it reusable
